refactor: Modularize movie details page and add trailer modal component. lint fix
