+++ src/app/movie/[id]/page.tsx
@@ -6,8 +6,8 @@ import { getMovieDetails } from "~/utils/api/tmdb";
 export async function generateMetadata({ params }: { params: { id: string } }) {
   const movie = await getMovieDetails(params.id);
   return {
-    title: `${movie?.title || "Not Found"} | Watchly`,
-    description: movie?.overview || "Not Found",
+    title: `${movie?.title ?? "Not Found"} | Watchly`,
+    description: movie?.overview ?? "Not Found",
   };
 }
