+++ src/components/global/modals/index.tsx
@@ -1,0 +1,3 @@
+import { default as TrailerModal } from "./trailer";
+
+export { TrailerModal };
