+++ src/components/resource/items/section/index.tsx
@@ -1,0 +1,3 @@
+import { default as ItemOverview } from "./itemOverview";
+
+export { ItemOverview };
